+++ torch/csrc/jit/codegen/cuda/scheduler/reduction.cpp
@@ -568,7 +568,7 @@ std::shared_ptr<ReductionParams> outerReductionHeuristic(
           // There's a place to put it in the device
           || target_blocks < device_multiprocessor_count * 4
           // There's a place to put it in unrolling
-          || target_unroll < vectorize_factor)) {
+          || target_unroll < int64_t(vectorize_factor))) {
     if (target_threads_in_block <
         ceilDiv(device_max_threads_per_multiprocessor, (int64_t)4)) {
       target_threads_in_block *= 2;
@@ -584,7 +584,7 @@ std::shared_ptr<ReductionParams> outerReductionHeuristic(
     if (target_blocks > device_multiprocessor_count &&
         target_threads_in_block >
             ceilDiv(device_max_threads_per_multiprocessor, (int64_t)16) &&
-        target_unroll < vectorize_factor && available_parallelism() > 1) {
+        target_unroll < int64_t(vectorize_factor) && available_parallelism() > 1) {
       target_unroll *= 2;
     }
   }
@@ -668,7 +668,7 @@ std::shared_ptr<ReductionParams> outerReductionHeuristic(
     iter_unroll_factor = std::min(iter_unroll_factor, iDimAvail());
     iter_unroll_factor = std::min(iter_unroll_factor, target_unroll);
     iter_unroll_factor = scheduler_utils::lastPow2(iter_unroll_factor);
-    if (vectorize_factor > 1 && iter_unroll_factor <= vectorize_factor) {
+    if (vectorize_factor > 1 && iter_unroll_factor <= (int64_t)vectorize_factor) {
       iter_unroll_factor =
           std::min(iter_unroll_factor, (int64_t)vectorize_factor);
       vectorize = true;
